Initial stab at DUALSHOCK 2 emulation
Poorly fakes a DUALSHOCK 2 for PS2 games that require it (hold UP on the D-pad on startup to use DS2 mode).  Analog sticks never work, and there's no way to press L3/R3 currently, though the definitions of them have been added, so they can be used in future mappings.
Also modified the mappings so that other button combos can be used on startup to activate other modes.